Split WinEnv into separate classes.   For ease of reuse and customization as a library   without wrapping.   WinEnvThreads is a class for replacement.   WintEnvIO is a class for reuse and behavior override.   Added private virtual functions for custom override   of fallocate pread for io classes.
