+++ port/win/port_win.h
@@ -69,7 +69,6 @@ typedef SSIZE_T ssize_t;
 namespace rocksdb {
 
 #define PREFETCH(addr, rw, locality)
-std::string GetWindowsErrSz(DWORD err);
 
 namespace port {
